requirements.txt
transformers
datasets
Pillow
torch
onnx
opencv-python
pycocotools